Arrays are now passed by reference but still need to be specified size in param, fixed errors with classes, added some required tests

+++ src/tools.py
@@ -12,7 +12,7 @@ semanticCube = {
         INT: {
             '+':INT,
             '-': INT,
-            '*': INT,
+            '*': FLOAT,
             '/': INT,
             '%': INT,
             '>': INT,
@@ -24,7 +24,7 @@ semanticCube = {
             '!=': INT,
             '+=': INT,
             '-=': INT,
-            '*=': INT,
+            '*=': FLOAT,
             '/=': INT
         },
         FLOAT: {
